requirements.txt
boto3==1.34.76
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
openai==1.30.1
google-generativeai==0.5.2
qdrant-client==1.8.0
streamlit==1.34.0
requests==2.32.2
